✨ feat(download): apply file name of dragged in/opened file to the note title
To better sum up, when the user opens a file with Ctrl+O or the command menu option, OR drags a file in (.txt or .md), the file name gets stored in `useState` and is able to be updated into the `input` area automatically.

Bare in mind, this only works until the user exits or refreshes the page, to which the initial file name is lost and the `input` field becomes empty as normal.