♻️ refactor(markdown)!: move markdown preview directly to editor and move to IndexedDB w/ Dexie
As the title says, markdown previews are now directly in the editor itself. No more seperate page for it. Toggle in the same way you previewed Markdown before.

💥 **BREAKING:**
This commit also moves away from localStorage and switches to using IndexedDB w/ Dexie for storing notes, re-writing most of the handling of the notes to use IndexedDB.

-----
This means anything (including notes) stored in localStorage will be lost after this change.
-----

Going forward, I probably won't change the behind-the-scenes of storing the notes anymore (or as much), as I've wanted to switch to IndexedDB since this is a note-taking application that can now handle multiple notes, which could all be very long in length. They could exceed localStorage's 5MB limit, so IndexedDB would be better for something like Notetxt.

fuck it this shouldn't brea i've tested a bunch but lets hope this doesn't break.