🧹 chore(preview): use more subtle indicator for markdown preview
Instead of a toast-like alert popping up at the bottom of the preview, potentially obstructing content, it now shows up on the same row as the title of the the note.

Now, it simply says 'Markdown' with a Markdown icon to indicate that you are previewing a note in Markdown.

Recommitted to add animations.