✨ feat: add slash (`/`) commands
Currently supports 11 different slash commands, which you can use by typing '/[command]' onto the beginning of a line in the Editor. You can view the full list of commands and their aliases in the README section. More may (?) be added as time goes on.

I also took the time to update most of the README. Was considering making docs externally, but sure right now. Changes include

- Updated & improved features list
- Listed keybinds/actions in a table
- Added 'Supported Slash (`/`) Commands' section