♻️ refactor(markdown): make markdown and editor side by side
You can now edit and preview your notes side by side-by-side with live Markdown rendering! No more toggling back and forth (as much). When you toggle the Markdown view now, you'll see both the Editor and Markdown view, so you don't need to go back and forth anymore.

Most of the Editor styling had to be rewritten for this, so it might be a little flunky and may have some edge cases (i hope not). Also, the scrollbars of the Editor and Markdown view are (mostly) in sync, but they will only sync when you scroll in the Markdown area.

Also updated some things to account for these changes.