✨ feat(editor): display a welcome note on first visit
Uses the database to track when a user has visited the Editor for the first time, and if true, shows the content from `intro.md`. For this welcome note ONLY, the Markdown preview mode will automatically be enabled in order to visualise the included checklist easier.

Existing notes/users are unaffected by this change, this is only for first-time visitors/users.