✨ feat(manager): add 'Note Summary' to note options dropdown
Thanks to this dropdown, I can now finally add this here because its much easier to add/remove stuff now, and I now have the space for it (unlike before, that's why it was not there before).

Same as the one from the Command Menu, but now accessible through the dropdown.