🧹 chore: update logo
Following tradition with all the changes / additions, the logo is now being updated as well! This may change again at some point, but this should do for now, as this one isn't stolen from Google.

Also updated the footer a little bit as it wasn't changed since I made the footer redesign.