🧹 chore: use a consistent date format
Was consolidating this and found that it's inconsistent all-round. This ensures a solidifed format of:

${month} ${day}, ${year}
(eg.) February 2, 2025

The note creation date is exempt from this. That will still show the suffix and time, as it's pretty useful there still. However, it's been cleaned up very slightly to seperate the date and time created better.