🐛 fix(command): do not display 'Recent Notes' section if there are no notes
ie. If 0 < length of recent notes, then the group won't show.

If vice versa (ie. > 0), then it will show as intended.

This is more so to handle when the page has just loaded and the section hasn't been populated yet due to the loading/fetching of notes from the DB.