✨ feat: add settings
notetxt finally has settings!

this should have been added much sooner but i was 1. too lazy, and 2. didn't know what settings to add. this is kinda scuffed but it gets the job done enough i guess.

this is accessible as a new button next to the 'note manager' button in the editor.

settings include:
- default note name
- default file type (when downloading)
- enable preview on new notes
- show recent notes in command menu
---
- delete all notes (was moved here! no longer in note manager)

more might(?) be added as time goes on.

also fixed some things like text color on buttons on other pages and etc