✨ feat: add note templates
Accessible from the Note Manager via a new button, promptly labelled "Templates".

This shows a dropdown allowing you to create a new template, featuring a template name and the content of the template. When selecting a template from the dropdown, a new note will automatically be created with the template content already inserted into the note.

---

In addition, the 'template content' textarea supports slash commands. In order to facilitate this, the existing logic was essentially moved into a hook, which allows it to be used in both the Editor and the template dialog, albeit a few things had to be changed (not in the hook, the surrounding code where it's used).

Also changed layout styling of note controls to allow them to go onto a second line (since it would go offscreen on mobile otherwise), along with other misc. things.