🐛 fix(markdown): fix scrollbars not properly syncing when scrolling in markdown preview
b85b213 added this synced scrolling stuff but it was using a method that didn't work too great, so the scrollbars would go out of sync almost instantaneously, causing one to end earlier than the other.

this fixes that, and the scrollbars now sync (when scrolling in the markdown preview).